+++ sgwt/library/__init__.py
@@ -6,7 +6,7 @@ Author: Luke Lowery (lukel@tamu.edu)
 File: sgwt/library/__init__.py
 Description: Library module initialization for built-in datasets.
 """
-from ..io import (
+from ..util import (
     get_cholmod_dll,
     
     # Kernels